+++ src/libraries/fiff/fiff_proj.cpp
@@ -107,7 +107,7 @@ void FiffProj::activate_projs(QList<FiffProj> &p_qListFiffProj)
     for(it = p_qListFiffProj.begin(); it != p_qListFiffProj.end(); ++it)
         it->active = true;
 
-    printf("\t%d projection items activated.\n", p_qListFiffProj.size());
+    printf("\t%d projection items activated.\n", (int)p_qListFiffProj.size());
 }
 
 //=============================================================================================================
